Move reference impl tests to another file
